+++ geotrek/trekking/static/trekking/points_reference.js
@@ -8,7 +8,7 @@ PointsReferenceField = L.GeometryField.extend({
 
     addTo: function (map) {
         L.GeometryField.prototype.addTo.apply(this, arguments);
-        var $toolbar = $(map.drawControl._container);
+        var $toolbar = $(this._drawControl._container);
         $toolbar.find('.leaflet-draw-draw-marker').attr('title', tr('Add a point of reference'));
         $toolbar.find('.leaflet-draw-edit-edit').attr('title', tr('Move points of reference'));
         $toolbar.find('.leaflet-draw-edit-remove').attr('title', tr('Delete a point of reference'));